prepareSession method extracted from MobileDormantRecyclerAdapter.kt and FixedRecyclerAdapter.kt to superclass
