+++ app/src/main/java/pl/llp/aircasting/screens/dashboard/fixed/FixedRecyclerAdapter.kt
@@ -25,23 +25,4 @@ class FixedRecyclerAdapter(
         return MyViewHolder(viewMvc)
     }
 
-    override fun prepareSession(session: Session, expanded: Boolean): Session {
-        if (!expanded) {
-            return session
-        }
-
-        var reloadedSession: Session? = null
-
-        runBlocking {
-            val query = GlobalScope.async(Dispatchers.IO) {
-                val dbSessionWithMeasurements = mSessionsViewModel.reloadSessionWithMeasurements(session.uuid)
-                dbSessionWithMeasurements?.let {
-                    reloadedSession = Session(dbSessionWithMeasurements)
-                }
-            }
-            query.await()
-        }
-
-        return reloadedSession ?: session
-    }
 }
